restructure azure storage issue #60
