+++ src/Bet.AspNetCore.AzureStorage/Bet.AspNetCore.AzureStorage.csproj
@@ -9,5 +9,13 @@
     <PackageTags>aspnetcore, dotnetcore, azure storage, azure blob, azure queue, azure table,</PackageTags>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" />
+    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Bet.Extensions.AzureStorage\Bet.Extensions.AzureStorage.csproj" />
+  </ItemGroup>
+
 
 </Project>
