+++ src/Bet.Extensions.AzureStorage/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Bet.AspNetCore.AzureStorage")]
